fixed bug in metric calc

+++ train_scripts/train_functions.py
@@ -132,7 +132,7 @@ def one_epoch_train(model, train_loader, optimizer, criterion, device, scaler, i
 
     total_loss /= iter_counter / iters_to_accumulate
     predictions = np.array(predictions, dtype=np.float)
-    ground_truth = np.array(ground_truth, dtype=np.float)
+    ground_truth = np.array(ground_truth, dtype=np.int)
     avg_auc, aucs, rocs_parameters = get_metric(predictions, ground_truth)
 
     return total_loss, avg_auc, aucs, rocs_parameters, (predictions, ground_truth), \
